requirements.txt
paste
webob
ipdb
ipython
iptools
redis
requests
redisco
python-slugify